requirements.txt
streamlit
ollama
json-log-formatter
openai
python-dotenv
pyyaml
